requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
openai==1.51.0
chromadb==0.5.15
sentence-transformers==3.1.1
numpy==1.26.4
httpx==0.27.2
